Update imperial_legion.json
File hasn't been maintained. Added the missing units and removed all the units which no longer exist/are no longer used.

Updated names in the format - FACTION - TYPE - TIER - ROLE

As the Legion unit documentation is out-of-date this file may require additional work. It's possible the wrong units are being called in some cases because there's no way to cross-check many of the unit files with the documentation available. Orbital factories not included yet as they're not in the unit doc.

The mod's filenames are a mess :p