Rework Tyr beam weapon efffect and hit.
Beam has had the shooty orbs removed as well as the main beam texture
becoming flat and a bit thinner.
I've added sparks that run the whole length of the beam to give it
that texture feeling plus playing on the spark like effect more.

I've added some subtle debris particles that float up from the ground,
and some not so subtle explosive rings to help communicate the aoe
range of the weapon.