Fix Viper not firing when chasing units
Fixes #596 (3) - this is a base game bug which is also seen in the Omega. Use of `attack_range_frac` alone caused a large change in trades with Avengers. Instead we implement a fake weapon in conjunction with `attack_range_frac` so that engagements still occur at full range. The values chosen are based on testing with chases against the Artemis, which requires a lower `attack_range_frac` than an ARKYD. It's possible that a unit with the right speed could also break this implementation, but it requires an engine fix to properly resolve.